agents/requirements.txt
langgraph
ipython
langchain
langchain-core
google-generativeai
langchain-google-genai
langchain_community
langchain-openai
duckduckgo-search
requests
beautifulsoup4
python-dotenv
gitpython
pygments
tree-sitter
tree-sitter-python
aiofiles
langchain-tavily
pygraphviz
pydot